+++ resources/ui-app/src/index.js
@@ -15,16 +15,37 @@ class <%= componentName %> extends React.Component {
   static propTypes = {
     match: PropTypes.object.isRequired,
     showSettings: PropTypes.bool,
+  };
+
+  constructor(props) {
+    super(props);
+
+    this.connectedExamplePage = props.stripes.connect(ExamplePage);
   }
 
   render() {
-    if (this.props.showSettings) {
+    const {
+      showSettings,
+      match: {
+        path
+      }
+    } = this.props;
+
+    if (showSettings) {
       return <Settings {...this.props} />;
     }
     return (
       <Switch>
-        <Route path={`${this.props.match.path}`} exact component={Application} />
-        <Route path={`${this.props.match.path}/examples`} exact component={ExamplePage} />
+        <Route
+          path={path}
+          exact
+          component={Application}
+        />
+        <Route
+          path={`${path}/examples`}
+          exact
+          component={this.connectedExamplePage}
+        />
       </Switch>
     );
   }